(WIP) reorganize and rename files
