+++ apps/api/src/domain/models/bibliographic-reference/book-bibliographic-reference/book-bibliographic-reference.entity.ts
@@ -1,3 +1,4 @@
+import { RegisterIndexScopedCommands } from '../../../../app/controllers/command/command-info/decorators/register-index-scoped-commands.decorator';
 import { DTO } from '../../../../types/DTO';
 import { ResultOrError } from '../../../../types/ResultOrError';
 import InvalidResourceDTOError from '../../../domainModelValidators/errors/InvalidResourceDTOError';
@@ -7,26 +8,27 @@ import { ResourceType } from '../../../types/ResourceType';
 import { isNullOrUndefined } from '../../../utilities/validation/is-null-or-undefined';
 import { Resource } from '../../resource.entity';
 import { IBibliographicReference } from '../interfaces/IBibliographicReference';
-import JournalArticleBibliographicReferenceData from './JournalArticleBibliographicReferenceData';
+import BookBibliographicReferenceData from './book-bibliographic-reference-data.entity';
 
-export class JournalArticleBibliographicReference
+@RegisterIndexScopedCommands([])
+export class BookBibliographicReference
     extends Resource
-    implements IBibliographicReference<JournalArticleBibliographicReferenceData>
+    implements IBibliographicReference<BookBibliographicReferenceData>
 {
     readonly type = ResourceType.bibliographicReference;
 
-    readonly data: JournalArticleBibliographicReferenceData;
+    readonly data: BookBibliographicReferenceData;
 
-    constructor(dto: DTO<JournalArticleBibliographicReference>) {
+    constructor(dto: DTO<BookBibliographicReference>) {
         super({ ...dto, type: ResourceType.bibliographicReference });
 
         if (isNullOrUndefined(dto)) return;
 
-        this.data = new JournalArticleBibliographicReferenceData(dto.data);
+        this.data = new BookBibliographicReferenceData(dto.data);
     }
 
     validateInvariants(): ResultOrError<Valid> {
-        const typeErrors = validateSimpleInvariants(JournalArticleBibliographicReference, this);
+        const typeErrors = validateSimpleInvariants(BookBibliographicReference, this);
 
         if (typeErrors.length > 0)
             return new InvalidResourceDTOError(this.type, this.id, typeErrors);
