+++ apps/api/src/domain/models/bibliographic-reference/common/bibliographic-reference-creator.entity.ts
