+++ apps/api/src/test-data/buildBibliographicReferenceTestData/buildJournalArticleBibliographicReferenceTestData.ts
@@ -1,5 +1,5 @@
-import { JournalArticleBibliographicReference } from '../../domain/models/bibliographic-reference/entities/journal-article-bibliographic-reference.entity';
 import { IBibliographicReference } from '../../domain/models/bibliographic-reference/interfaces/IBibliographicReference';
+import { JournalArticleBibliographicReference } from '../../domain/models/bibliographic-reference/journal-article-bibliographic-reference/journal-article-bibliographic-reference.entity';
 import { BibliographicReferenceType } from '../../domain/models/bibliographic-reference/types/BibliographicReferenceType';
 import { CreatorType } from '../../domain/models/bibliographic-reference/types/CreatorType';
 import { ResourceType } from '../../domain/types/ResourceType';
